django-rest-sample/requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.0.14
django-environ==0.4.5
djangorestframework==3.11.2
djangorestframework-simplejwt==4.4.0
idna==2.10
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.25.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.4
